Move HotSwap implementation into ROCR
