+++ projects/rocr-runtime/runtime/hsa-runtime/core/runtime/hotswap.hpp
@@ -1,0 +1,68 @@
+////////////////////////////////////////////////////////////////////////////////
+//
+// The University of Illinois/NCSA
+// Open Source License (NCSA)
+//
+// Copyright (c) 2026, Advanced Micro Devices, Inc. All rights reserved.
+//
+// Developed by:
+//
+//                 AMD Research and HSA Software Development
+//
+//                 Advanced Micro Devices, Inc.
+//
+//                 www.amd.com
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to
+// deal with the Software without restriction, including without limitation
+// the rights to use, copy, modify, merge, publish, distribute, sublicense,
+// and/or sell copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following conditions:
+//
+//  - Redistributions of source code must retain the above copyright notice,
+//    this list of conditions and the following disclaimers.
+//  - Redistributions in binary form must reproduce the above copyright
+//    notice, this list of conditions and the following disclaimers in the
+//    documentation and/or other materials provided with the distribution.
+//  - Neither the names of Advanced Micro Devices, Inc,
+//    nor the names of its contributors may be used to endorse or promote
+//    products derived from this Software without specific prior written
+//    permission.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+// THE CONTRIBUTORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+// OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+// ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+// DEALINGS WITH THE SOFTWARE.
+//
+////////////////////////////////////////////////////////////////////////////////
+
+#ifndef HSA_RUNTIME_CORE_RUNTIME_HOTSWAP_HPP_
+#define HSA_RUNTIME_CORE_RUNTIME_HOTSWAP_HPP_
+
+#include <cstdlib>
+#include <memory>
+
+#include "core/inc/amd_hsa_loader.hpp"
+#include "inc/hsa.h"
+
+namespace rocr {
+namespace hotswap {
+
+using OwnedElf = std::unique_ptr<void, decltype(&std::free)>;
+
+bool TryRetargetCodeObject(amd::hsa::loader::CodeObjectReaderImpl* reader,
+                           hsa_agent_t agent, OwnedElf* out_elf,
+                           size_t* out_elf_size);
+
+void RetainElf(hsa_executable_t executable, OwnedElf elf);
+void ReleaseElfs(hsa_executable_t executable);
+void LogRewrittenLoadFailure(hsa_status_t status);
+
+}  // namespace hotswap
+}  // namespace rocr
+
+#endif  // HSA_RUNTIME_CORE_RUNTIME_HOTSWAP_HPP_
