+++ projects/rocr-runtime/runtime/hsa-runtime/core/runtime/runtime.cpp
@@ -2811,14 +2811,7 @@ void Runtime::LoadTools() {
               getpid(), rocp_reg_status, rocprofiler_register_error_string(rocp_reg_status));
     }
 
-    // rocprofiler-register (v3) provides tracing only; the hotswap tool must
-    // intercept and modify HSA calls (it wraps hsa_code_object_reader_create_from_memory),
-    // which requires the v1 HSA_TOOLS_LIB path. Allow v1 registration specifically for
-    // that first-party tool so it loads via HSA_TOOLS_LIB alone, without re-enabling v1
-    // for other tools. General v1 behavior is otherwise unchanged.
-    static constexpr const char* kHotswapToolLib = "libhsa-hotswap.so";
-    bool allow_v1_registration =
-        flag().tools_lib_names().find(kHotswapToolLib) != std::string::npos;
+    bool allow_v1_registration = false;
     if (os::IsEnvVarSet("HSA_TOOLS_ROCPROFILER_V1_TOOLS")) {
       // assume true if env variable is set
       allow_v1_registration = true;
